fixed most comments
