+++ src/main/java/com/epam/esm/controller/TagController.java
@@ -18,6 +18,13 @@ public class TagController {
     @Autowired
     private TagService tagService;
 
+    /**
+     * Gets all tags
+     * @param params MVM of strings to get all tags
+     * @return List of tags
+     * @throws DaoException
+     */
+
     @RequestMapping(method = RequestMethod.GET)
     public List<Tag> findAll(@RequestParam(required = false) MultiValueMap<String, String> params) throws DaoException {
         try {
@@ -26,6 +33,13 @@ public class TagController {
             throw new DaoException(e);
         }
     }
+
+    /**
+     * Adds a new tag to list
+     * @param tag
+     * @return tag object
+     * @throws DaoException
+     */
 
     @RequestMapping(method = RequestMethod.POST)
     public Tag add(@RequestBody @Validated Tag tag) throws DaoException {
@@ -36,7 +50,14 @@ public class TagController {
         }
     }
 
-    @RequestMapping(method = RequestMethod.DELETE)
+    /**
+     * Deletes a tag
+     * @param tag tag object
+     * @param id an to perform an operation by
+     * @throws DaoException
+     */
+
+    @RequestMapping(method = RequestMethod.DELETE, value = "/{id}")
     public void delete(@RequestBody Tag tag, @PathVariable Long id) throws DaoException {
         try {
             tagService.delete(tag, id);
