--- src/main/java/com/epam/esm/dao/RowMapProvider.java
@@ -1,13 +1,0 @@
-package com.epam.esm.dao;
-
-public class RowMapProvider {
-
-    public static final String PROVIDER_ID = "id";
-    public static final String PROVIDER_NAME = "name";
-    public static final String PROVIDER_DESCRIPTION = "description";
-    public static final String PROVIDER_PRICE = "price";
-    public static final String PROVIDER_DURATION= "duration";
-    public static final String PROVIDER_CREATE_DATE = "create_date";
-    public static final String PROVIDER_LAST_UPDATE_DATE = "last_update_date";
-
-}
