+++ src/main/java/com/epam/esm/service/TagService.java
@@ -1,8 +1,14 @@
 package com.epam.esm.service;
 
+import com.epam.esm.dao.DaoException;
 import com.epam.esm.model.Tag;
+import org.springframework.util.MultiValueMap;
 
+import java.util.List;
+
+/**
+ * Interface for TagService entity
+ */
 
 public interface TagService extends CrudService<Tag> {
-
 }
