+++ src/main/java/com/epam/esm/service/impl/TagServiceImpl.java
@@ -8,11 +8,14 @@ import com.epam.esm.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
 import org.springframework.util.MultiValueMap;
 
 
 import java.util.List;
 @Service
+@Transactional(readOnly = true)
 public class TagServiceImpl implements TagService {
 
     private TagDao tagDao;
@@ -24,12 +27,15 @@ public class TagServiceImpl implements TagService {
     @Override
     public List<Tag> getByParameters(MultiValueMap<String, String> params) throws DaoException {
         try {
-            return tagDao.getAll();
+            if (CollectionUtils.isEmpty(params))
+                return tagDao.getAll();
+            return tagDao.getByParameters(params);
         } catch (DataAccessException e) {
             throw new DaoException(e);
         }
     }
 
+    @Transactional
     @Override
     public Tag add(Tag tag) throws DaoException {
         try {
@@ -39,9 +45,16 @@ public class TagServiceImpl implements TagService {
         }
     }
 
+    @Transactional
     @Override
     public void delete(Tag tag, Long id) throws CertificateServiceException, DaoException {
-
+        try {
+            tagDao.getById(id);
+            tag.setId(id);
+            tagDao.delete(tag);
+        } catch (DaoException e) {
+            throw new CertificateServiceException(e);
+        }
     }
 
     @Override
